# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Shoe Store API",
    "version": "1.0.0",
    "description": "API for managing a shoe store"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local server"
    }
  ],
  "components": {
    "schemas": {
     "Stock": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "size": {
            "type": "number"
          },
          "quantity": {
            "type": "number"
          }
        },
        "required": ["id", "size", "quantity"]
      },
      "Shoe": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "brand": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "stocks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Stock"
            }
          },
          "colors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "imageUrl": {
            "type": "string"
          },
          "descriptions": {
            "type": "string"
          }
        },
        "required": ["name", "brand", "price", "stocks", "colors", "imageUrl","descriptions"]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "gender": {
            "type": "string"
          },
          "dob": {
            "type": "string",
            "format": "date"
          },
          "imageUrl": {
            "type": "string"
          }
        },
        "required": ["name", "email", "password"]
      },
...
    "CartItem": {
        "type": "object",
        "properties": {
          "shoeId": {
            "type": "string"
          },
          "stock": {
            "$ref": "#/components/schemas/Stock"
          }
        },
        "required": ["shoeId", "stock"]
      },
      "Cart": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CartItem"
            }
          }
        },
        "required": ["userId", "items"]
      },
...
    }
  },
  "paths": {
  "/shoes": {
        "post": {
          "summary": "Create a new shoe",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Shoe"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Shoe created successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Shoe"
                  }
                }
              }
            }
          }
        },
        "get": {
          "summary": "Get all shoes",
          "responses": {
            "200": {
              "description": "A list of shoes",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Shoe"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/shoes/{id}": {
        "get": {
          "summary": "Get a single shoe by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "A single shoe",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Shoe"
                  }
                }
              }
            },
            "404": {
              "description": "Shoe not found"
            }
          }
        },
        "put": {
          "summary": "Update a shoe by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Shoe"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Shoe updated successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Shoe"
                  }
                }
              }
            },
            "404": {
              "description": "Shoe not found"
            }
          }
        },
        "delete": {
          "summary": "Delete a shoe by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "204": {
              "description": "Shoe deleted successfully"
            },
            "404": {
              "description": "Shoe not found"
            }
          }
        }
      },
...
  "/carts": {
      "post": {
        "summary": "Create a new cart",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Cart"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Cart created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cart"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get all carts",
        "responses": {
          "200": {
            "description": "A list of carts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Cart"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/carts/{id}": {
      "get": {
        "summary": "Get a single cart by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A single cart",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cart"
                }
              }
            }
          },
          "404": {
            "description": "Cart not found"
          }
        }
      },
      "/cart/item/{id}": {
      "put": {
        "summary": "Update a cart by ID cart",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Cart"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cart updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cart"
                }
              }
            }
          },
          "404": {
            "description": "Cart not found"
          }
        }
      }
    } ,
      "delete": {
        "summary": "Delete a cart by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Cart deleted successfully"
          },
          "404": {
            "description": "Cart not found"
          }
        }
      }
    },
...
    "/users/{id}": {
      "get": {
        "summary": "Get user details by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User details retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "put": {
        "summary": "Update user details by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    }
  }
 }